Make ShorthandManager to be a separate class
